renamed some aspects, removed importing abstract classes in __init__
