+++ credoai/artifacts/__init__.py
@@ -1,4 +1,2 @@
 from .data.tabular_data import TabularData
-from .data.data import Data
 from .model.classification_model import ClassificationModel, DummyClassifier
-from .model.model import Model
